requirements.txt
packaging == 22.0
Pillow == 9.3.0
pip == 22.3.1
pytesseract == 0.3.10
setuptools == 57.4.0
